--- axelar-contracts/constants/addresses.json
@@ -1,5 +1,0 @@
-{
-    "avalanche": ["0x497221063544fb04bf855742448f2207e79d52c4"],
-    "fantom": ["0xd51a2d698137e2f70c54fd626678fcaa66b2e69d"],
-    "polygon": ["0xda77169fe441173b190590e568272d7deeee2f82"]
-}
