+++ axelar-contracts/package.json
@@ -27,6 +27,7 @@
         "bip39": "^3.0.4",
         "dotenv": "^16.0.2",
         "ethers": "^5.6.2",
+        "nft.storage": "^7.0.3",
         "uuid": "^8.3.2"
     },
     "devDependencies": {
